fix display error message, create constants.js,validation.js
