--- src/index.js
@@ -1,7 +1,0 @@
-import {
-  CustomerManagerModel,
-  CustomerView,
-  CustomerController
-} from './modules/customer/index.js'
-
-const customerApp = new CustomerController(new CustomerManagerModel (), new CustomerView ())
