+++ src/js/index.js
@@ -1,0 +1,7 @@
+import {
+  CustomerModel,
+  CustomerView,
+  CustomerController
+} from './modules/customer/index.js'
+
+const customerApp = new CustomerController(new CustomerModel (), new CustomerView ())
