+++ src/js/modules/customer/customer.model.js
@@ -1,0 +1,8 @@
+import {HttpService} from '../../service/service.js'
+import { customerApi } from '../../constants/constants.js'
+
+export class CustomerModel {
+  async addCustomer(data) {
+    await HttpService.post(customerApi,data)
+  }
+}
