+++ src/js/modules/customer/index.js
