+++ src/js/service/service.js
@@ -1,4 +1,4 @@
-export const Service = {
+export const HttpService = {
   async post(path, data) {
     await fetch(path, {
       method: 'POST',
